basic periodic polling of market book + re-work
I've taken the approach of having a single connection
per market request periodically polling. Will probably change
this to per event or some other grouping. Still need to work
out how this lib will be used from a user perspective.

Decided to remove all the session management stuff which i
had tried to hide from the user behind a simpler api but this was
proving too difficult. To keep things simple and to the point
made session management a user concern, in that they are required
to obtain the user session token, pass it in on requests and
eventually to keep it alive, all via the api.

Opted to go with the built in erlang httpc lib for all requests
other than list_market_book as they will be executed
infrequently and so don't require a persistent connection. Only
list_market_book requires this, so i've stuck with gun for that.